requirements.txt
aiohappyeyeballs
aiohttp
aiohttp-retry
aiosignal
arabic-reshaper
arrow
asgiref
asn1crypto
attrs
beautifulsoup4
binaryornot
certifi
cffi
chardet
charset-normalizer
click
colorama
cookiecutter
crispy-bootstrap5
crispy-tailwind
cryptography
cssselect2
distlib
dj-database-url
Django
django-allauth
django-appconf
django-bootstrap-v5
django-browser-reload
django-compressor
django-cors-headers
django-crispy-forms
django-debug-toolbar
django-environ
django-filter
django-mysql
django-tailwind
django-tailwindcss
djangorestframework
djangorestframework-simplejwt
filelock
frozenlist
git-filter-repo
html5lib
idna
Jinja2
lxml
Markdown
markdown-it-py
MarkupSafe
mdurl
MouseInfo
multidict
multipledispatch
mysql-connector-python
numpy
opencv-contrib-python
opencv-python
oscrypto
pillow
pip
pipenv
platformdirs
propcache
psycopg2
psycopg2-binary
PyAutoGUI
pycparser
PyGetWindow
Pygments
pyHanko
pyhanko-certvalidator
PyJWT
PyMsgBox
pypdf
pyperclip
PyRect
PyScreeze
python-barcode
python-bidi
python-dateutil
python-slugify
pytweening
PyYAML
qrcode
rcssmin
reportlab
requests
rich
rjsmin
setuptools
six
sortedcontainers
soupsieve
sqlparse
sslcommerz-lib
svglib
text-unidecode
tinycss2
twilio
types-python-dateutil
typing_extensions
tzdata
tzlocal
uritools
urllib3
virtualenv
webencodings
wheel
xhtml2pdf
yarl
